requirements.txt
garminconnect==0.2.8
python-dotenv==0.21.0
pytz==2022.5